+++ src/main/java/dto/ExchangeDTO.java
@@ -1,23 +1,25 @@
 package dto;
 
+import java.math.BigDecimal;
+
 public class ExchangeDTO extends ExchangeRateDTO {
 
-    private final double amount;
-    private final double convertedAmount;
+    private final BigDecimal amount;
+    private final BigDecimal convertedAmount;
 
     public ExchangeDTO(int id, CurrencyDTO baseCurrencyDto,
-                       CurrencyDTO targetCurrencyDto, double rate,
-                       double amount, double convertedAmount) {
+                       CurrencyDTO targetCurrencyDto, BigDecimal rate,
+                       BigDecimal amount, BigDecimal convertedAmount) {
         super(id, baseCurrencyDto, targetCurrencyDto, rate);
         this.amount = amount;
         this.convertedAmount = convertedAmount;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public double getConvertedAmount() {
+    public BigDecimal getConvertedAmount() {
         return convertedAmount;
     }
